Move Result to file
